pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sm-model-governance"
version = "0.2.0"
description = "Three-plane ML governance with cryptographic approvals for NANDA agent registries"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["nanda", "agent", "model", "governance", "approval", "drift", "ml-ops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = []

[project.optional-dependencies]
crypto = ["cryptography>=41.0"]
postgres = ["psycopg2-binary>=2.9"]
integrity = ["sm-model-integrity-layer>=0.1.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "ruff>=0.1.0", "mypy>=1.0", "coverage>=7.0"]
all = ["sm-model-governance[crypto,postgres,integrity]"]

[tool.hatch.build.targets.wheel]
packages = ["sm_governance"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
